+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/constants/AbfsHttpConstants.java
@@ -30,7 +30,11 @@ import org.apache.hadoop.util.VersionInfo;
 public final class AbfsHttpConstants {
   // Abfs Http client constants
   public static final String FILESYSTEM = "filesystem";
+  public static final String CONTAINER = "container";
+  public static final String METADATA = "metadata";
+  public static final String LIST = "list";
   public static final String FILE = "file";
+  public static final String LEASE = "lease";
   public static final String DIRECTORY = "directory";
   public static final String APPEND_ACTION = "append";
   public static final String FLUSH_ACTION = "flush";
@@ -46,6 +50,7 @@ public final class AbfsHttpConstants {
   public static final String DEFAULT_LEASE_BREAK_PERIOD = "0";
   public static final String DEFAULT_TIMEOUT = "90";
   public static final String APPEND_BLOB_TYPE = "appendblob";
+  public static final String BLOCK_BLOB_TYPE = "BlockBlob";
   public static final String TOKEN_VERSION = "2";
 
   public static final String JAVA_VENDOR = "java.vendor";
@@ -88,6 +93,7 @@ public final class AbfsHttpConstants {
   public static final String HTTP_HEADER_PREFIX = "x-ms-";
   public static final String HASH = "#";
   public static final String TRUE = "true";
+  public static final String ZERO = "0";
 
   public static final String PLUS_ENCODE = "%20";
   public static final String FORWARD_SLASH_ENCODE = "%2F";
